+++ src/noc/containers/noc-field/field-type-enum.model.ts
