+++ src/noc/containers/noc-field/noc-field.component.ts
