+++ src/noc/containers/noc-field/palette.service.ts
