+++ src/noc/containers/noc-field/utils/noc-validators.ts
