+++ src/noc/containers/noc-field/yes-no/index.ts
