+++ src/noc/containers/noc-field/yes-no/noc-yes-no-field.html
