+++ src/noc/containers/noc-field/yes-no/yes-no.service.ts
