+++ src/noc/containers/noc-q-and-a/noc-q-and-a.component.html
@@ -7,13 +7,13 @@
 
   <p>You must enter the client details exactly as they’re written on the case, including any mistakes.</p>
 
-  <form class="form" [formGroup]="qAndAForm">
+  <form class="form" [formGroup]="formGroup">
     <fieldset id="fieldset-q-and-a-form">
       <ng-container *ngFor="let question of questions$ | async">
         <div>
           <ng-container>
             <exui-noc-field [questionField]="question"
-                            [formGroup]="qAndAForm"
+                            [formGroup]="formGroup"
                             [idPrefix]=""></exui-noc-field>
           </ng-container>
         </div>
