+++ src/noc/noc.module.ts
@@ -9,14 +9,14 @@ import { SharedModule } from 'src/app/shared/shared.module';
 import { AppConfig } from '../app/services/ccd-config/ccd-case.config';
 // from containers
 import * as fromContainers from './containers';
-import { NocEmailFieldComponent } from './containers/noc-filed/email';
-import { FormValidatorsService } from './containers/noc-filed/form-validators.service';
-import { NocNumberFieldComponent } from './containers/noc-filed/number';
-import { PaletteService } from './containers/noc-filed/palette.service';
-import { NocPhoneUkFieldComponent } from './containers/noc-filed/phone-uk';
-import { NocPostcodeFieldComponent } from './containers/noc-filed/postcode';
-import { NocTextFieldComponent } from './containers/noc-filed/text';
-import { NocYesNoFieldComponent, YesNoService } from './containers/noc-filed/yes-no';
+import { NocEmailFieldComponent } from './containers/noc-field/email';
+import { FormValidatorsService } from './containers/noc-field/form-validators.service';
+import { NocNumberFieldComponent } from './containers/noc-field/number';
+import { PaletteService } from './containers/noc-field/palette.service';
+import { NocPhoneUkFieldComponent } from './containers/noc-field/phone-uk';
+import { NocPostcodeFieldComponent } from './containers/noc-field/postcode';
+import { NocTextFieldComponent } from './containers/noc-field/text';
+import { NocYesNoFieldComponent, YesNoService } from './containers/noc-field/yes-no';
 import { nocRouting } from './noc.routes';
 import { NocService } from './services';
 import { effects, reducers } from './store';
